+++ core/src/ten_utils/backtrace/platform/posix/linux/elf_internal/crc32.c
@@ -7,7 +7,7 @@
 // This file is modified from
 // https://github.com/ianlancetaylor/libbacktrace [BSD license]
 //
-#include "include_internal/ten_utils/backtrace/platform/posix/linux/crc32.h"
+#include "include_internal/ten_utils/backtrace/platform/posix/linux/elf_internal/crc32.h"
 
 #include <errno.h>
 #include <sys/stat.h>
